fix(lint): move wizard helpers to utils module
