+++ src/views/admin/TournamentNewWizard.jsx
@@ -2,6 +2,8 @@ import React, { useMemo, useState } from "react";
 import PropTypes from "prop-types";
 import "./tournamentNewWizard.css";
 
+import { FRANCHISE_COLOUR_ROTATION, normaliseId } from "./TournamentNewWizard.utils";
+
 const STEPS = [
   "Tournament Details",
   "Franchises",
@@ -10,27 +12,7 @@ const STEPS = [
   "Review & Submit",
 ];
 
-export const FRANCHISE_COLOUR_ROTATION = [
-  "#2E5BFF",
-  "#22C55E",
-  "#F97316",
-  "#A855F7",
-  "#EF4444",
-  "#06B6D4",
-  "#F59E0B",
-  "#10B981",
-  "#6366F1",
-  "#EC4899",
-];
 
-export function normaliseId(name) {
-  return String(name || "")
-    .trim()
-    .toLowerCase()
-    .replace(/[^a-z0-9]+/g, "-")
-    .replace(/(^-|-$)/g, "")
-    .slice(0, 64);
-}
 
 const franchiseShape = PropTypes.shape({
   id: PropTypes.string.isRequired,
